child-rekey: Suppress updown event when deleting redundant CHILD_SAs
When handling a rekey collision we might have to delete an already
installed redundant CHILD_SA (or expect the other peer to do so).  We don't
want to trigger updown events for these as we don't during rekeying.

Instead of setting the state to CHILD_REKEYING we could maybe use
CHILD_REKEYED, which we currently only use for IKEv1, and set it for
all CHILD_SAs we delete or expect the other peer to delete.  Would need
a small change in child-delete too.  Or we could introduce a new state.

 #853.